Improve column-order tests: remove pointless tests, add meaningful ones, link guide
Split the 452-line column-order.test.ts into focused files:
- test/lib/column-order/template.test.ts — template parsing, validation, rendering
- test/lib/column-order/columns.test.ts — event and attendee column cell renderers

Removed 17 pointless tests that were tautological or trivial property accessors
(e.g. "location returns location", "rawValue returns property"). Added 8 useful
tests: end-to-end renderCells pipeline, Liquid filter integration, XSS escaping,
CSS class application, getHeaderText fallback, status/attendees cell logic.

Also links the column order settings forms to the admin guide's Column Order
section so users can discover filter syntax while editing their template.
